feat!: move temporal load balancer to private by default
The Temporal LB only needs to be public to allow agent connections.
Since self-hosted installs do not use agent, the Temporal service
can be internal and not exposed to the public internet.

BREAKING CHANGE: A new variable `temporal_internet_facing`
has been introduced to control whether the Temporal LB is internet
facing. The default is `false`, which is a breaking change causing
the LB to be destroyed and recreated.

Recommendation: accept the new default and migrate to an internal
temporal LB. This is more secure since it avoids unnecessary public
access to the Temporal LB.

Downtime: Yes. There will be a service interruption for this change.
While the service is offline, workers will not be able to retrieve work
and no new work (i.e. metric runs) will be able to be published to the
work queue. Work already in queue will remain there and get picked up
as soon as the Temporal service is restored.

Steps: Upgrade the terraform module version and run terraform apply.
Terraform will destroy and recreate the load balancer if necessary.
For customers who do not wish to make this change, or wish to make this
change at a later date, set the `temporal_internet_facing`
variable to `true`.