feat: add VPC endpoints for resources required for ECS
This will help keep important traffic off of public net (like ASM)
as well as reduce ingress and egress costs for data transfers
(like image pulls and ML models stored in s3).

There is a ~$8/mo per AZ charge for VPCe's so I opted not to add
ones that are there just for the admin container debugging.  It's
not critical for that traffic to stay off of private net anyhow.