feat: add BYO mail server env vars to datawatch, datawork, metricwork
These vars are for plumbing in a custom SMTP server.  The app is
set up to fall back to default for both empty string env vars
and non-existant env vars.

I opted to only enable the feature if all 4 required env vars are
specified to avoid any kind of mistakes when trying to use the feature