fix: always ACM cert creation flag logic
The logic should be:

Always create the ACM cert if it isn't passed in.

So it shouldn't rely on DNS control at all since Bigeye requires
an ACM cert for SSL on the LBs as we want to enforce HTTPS
every where.

Right now the logic does not cover the case where nothing is set
(ie all defaults) which is a bug introduced in 1.14.0