docs: update self-managed-vpc-no-internet-access instructions
There are tons of new features now, update this set of instructions
with a working end to end install.  Normal installs will only require
what's in main.tf, but for the purpose of creating a functional
end to end install, all of the addtional *_example.tf files have been
added to create addtional resources necessary such as SES, Workmail,
a bastion for SSH tunnel, etc.