docs: fix bug with no-internet vpc example and ECR privatelink
A couple things in this PR that are subtle.

1. There is what I would call a bug with ECR privatelink that both
an S3 gateway and S3 interface endpoint must be created or ECR image
pulls will fail due to some portion of the request chain going out
over public net (even though we don't even have a route to public net).

This is solved in our example TF by adding a S3 *Interface* VPCE
and it specifically must have private_dns_only_for_inbound_resolver_endpoint
set to false or the requests still end up resolving with a public
IP and thus failing.

2. The terraform-aws-modules/vpc/aws//modules/vpc-endpoints version
that we were using had a bug where private_dns_only_for_inbound_resolver_endpoint
was not getting applied to the VPCE.  Updating to > 5.5.0 fixes
that problem as well.

With these changes, the example is now able to work with security
group rules completely restricted to just the VPC CIDR for all
security group ingress and egress rules.