fix: add app tag for temporal, redis, rabbit
The app tag is nice for billing/search/etc and we should apply it
to all resources related to services.  The one oddity will be
shared resources that several services use.  We have been using
datawatch as the app name for those historically, so should continue
to do so for consistency.