fix: set default for performance insights back to true
RDS Performance insights are inexpensive enough that it is recommended
this is on for all databaess (where the hardware supports it).  The
change in 3.11.1 fixed the logic for enabling only if the hardware
is not in the blacklist, but ended up disabling performance
monitoring by default. This fix turns it back on by default.