feat: add lineagework service
This is splitting out the lineage background processes into its
own dedicated set of workers to improve service/job isolation and
improve visibility.

While not a breaking change, installs that do not have lineage plus
installed can just set the lineagework instance count to 0 to save
on costs as that service only pertains to work for the lineage plus
feature.  Installs with lineageplus will want to tune cpu/mem
for both the datawork and lineagework services after deployment.
Setting CPU/mem for lineagework to the same as datawork is a reasonable
starting point.