fix: rename papi service to internalapi
papi (private api) ended up not being a popular name internally,
so it is being renamed to internalapi.  It was a thought to
support the existing variable names and avoid the breaking change,
but it will keep the house much cleaner if we just take the
breaking change and remove them.

BREAKING CHANGE: All variables with papi in the name need to
be globally replaced with internalapi.