feat: use an include queue list instead of exclude queue
There is a new feature in datawatch and related services to allow
an include list.  Using this removes the maintenance task of needing
to maintain the exclude list everytime a new queue is added to temporal.

Also convert metricwork over to use this instead of the single queue
override.  That feature is (or should be) deprecated and removed
since we now have an include list.