feat: add indexwork service
This service will be added in this version, but will be enabled to pick
up work behind a flag so we can roll it out safely on a per stack basis.

This service will be dedicated to handle the heavy catalog indexing
queue to try and isolate the work and improve memory management and
instance overloading in large installations.