feat!: rollout indexwork service
Catalog indexing is a CPU intensive process that is quite bursty
in nature.  As such it has been moved to a dedicated service for
resource isolation and hardware efficiency.  Note that the service
will autoscale down to 0 instances when the queue is empty.  This
is normal.

BREAKING CHANGE: The following feature flags will need to be removed
from your config if you are using them:

- indexwork_enabled
- indexwork_autoscaling_enabled