feat: scale indexwork in faster
Scale out ignores the cooldown so will happen whenever a new
message shows up on the queue, but we can scale-in faster by having
a shorter cooldown.

After some observation we have quite a few periods where there are
very small numbers of jobs on the queue that trickle in, so scaling
in faster will be a little more efficient.

This time is still longer than the instance stop delay so will
not affect how long we wait for a task to terminate.