feat: enable auto minor version upgrade for rabbitmq
Starting with 3.13, AWS requires auto minor version upgrade to
be enabled.  This also has to be set _before_ trying to upgrade
to 3.13.  So it is just best to set it to true now.

If someone has not touched the default already, leave it set to
false.  It will get turned on when they upgrade to 3.12.x anyways
and that is required before they upgrade to 3.13 as it is not
possible to upgrade from 3.11.x -> 3.13 directly.