feat: enable datawatch_request_body_logging_enabled everywhere
This flag just controls weather a trace id is logged in the response
log from the backend server.  It will be set on (flag control removed)
in Bigeye app version 2.2.0 anyways, so we are enabling this across
the board in the terraform ahead of the code being released.

This variable will be removed in a future release.