fix: update dns to only route static assets to cloudfront (#425)
* fix: update dns to only route static assets to cloudfront

The current implementation routed all requests through CF.  For
API calls in particular this can be a problem as CF has rather short
timeouts for API requests.

This creates a separate URL that the app can use just for requesting
static assets from Cloudfront.

Other updates in this commit:
- introduce a 2nd flag to allow enable/disable of Cloudfront without a brief
outage due to a race between DNS changes and cloudfront distribution
creation/deletion
- allow setting Cloudfront request timeout
- set Cloudfront changes to wait for deploy.  This avoids a race with
DNS changes when enabling/disabling cloudfront that will lead to
downtime (routing to Cloudfront before it has been deployed)

* fix: add env var to control static asset address for web

We will want to tell the web service where to go for static assets
as the TF side is what is aware of whether traffic should be routed
to cloudfront and also where to find it.