feat: add a dedicated general purpose internal load balancer
No actual services are being migrated through the new ALB in this
PR.

This is the first step to consolidating the internal load balancers
into 1 and using path based routing instead of load balancer per
service.

This is a cost savings step as each LB is ~$23/mo and it adds up
with so many internal services.

Future PRs will migrate the services over and remove the dedicated
per-service ALBs.