feat!: add dedicated lineage api service
This adds a new service - lineageapi.  Its purpose will be to
service backend calls for lineage related endpoints.

BREAKING CHANGE: as part of adding a new service, a previous change
to security group rules had to be cleaned up.  No action by installers
is required other than to install and upgrade to the previous major
version first before installing this one.

Note that this PR does not handle routing of API calls, that will
be an app version change (TBD).  This PR only handles installation
of the service.