feat!: migrate off of inline security group rules part 1/2
This is the first step in a 2 part series to migrate from inline
security group rules to dedicated rules.  The inline rules do
not track AWS rule Ids properly which blocks seamless changes.

Part 1 sets the inline rule to an empty list.  Without this,
the old rules are left dangling and will conflict with new rules.

Part 2 is to remove the inline rules completely to avoid the inline
rules fighting with the dedicated ingress/egress rules for control

BREAKING CHANGE: Part one should only be applied 1x.  Then upgrade
to part 2 and the plan will be idempotent again.