fix: pin sha for all github actions
The test and publish action is failing, might as well handle this
now since the actions all have to be tested anyways.

Pinning the sha is a security enhancement over using tagged versions
as the tags are not immutable and can be retagged onto different
commits.  Most deterministic is using the SHA (this is GH's
recommendation as well)