feat: allow using fargate spot instances
By default no spot instances will be used.  This PR will allow
use of spots for cost savings to leverage spot instances for ECS
where possible without impacting reliability.  Spot instances
are not reclaimed often in ECS and when they are it is typically
in a single AZ, so instances will simply redeploy to a different
AZ.

There is one setting across the board vs per service settings to
keep things tidy.

NOTE that the solr service for LineagePlus is ommitted from spot
instances as it does not tolerate node re-deployment without
an interruption in service.