feat!: allow enabling alb anomaly mitigation
Anomaly mitigation is a newer routing feature on ALB which ideally
solves both hot spots due to large requests as well as avoiding the
case where a sick instance (like with a broken DB connection) can
get a large percentage of inbound requests routed to it because
it processes them very quickly and thus has the "least number of
active connections".

This isn't available in conjunction with session stickiness so
is not used on the frontend, but will set this everywhere else.

BREAKING CHANGE: The aws provider version upgrade is required in
order to control the enable flag for this feature.