fix: update ALB unhealthy check count for API servers
This is the number of consecutivly failing health checks before
the ALB marks a node as bad and terminutes it.

We are currently waiting 10 x 30 (5 minutes) before marking an instance
unhealthy which is a very long time for an API server to be
"unavailable", but still accepting requests.

2.5 minutes seems like a better balance to give servers the chance
to service requests and start responding the health checks.  Faster
might seem better, but may lead to nodes cycling out and not
completing work if we have computationally intense jobs come in,
which tends to happen in our environment.

Overall the goal here is to not wait around for some huge workload
that will never complete, ie fail faster, as the busy node will
have other work plugged up on it (innocent by standers).

Monocle also is basically just another API server so does not need
configuration different than our java API servers except that it
does not need a long warm up time (health check grace period)
since it doesn't have much in the way of caching that needs warming up.