fix: adjust spot instance configs
LB deregistration delay and instance stop time need to finish before
the 2 minute grace period that is given when spot instances are
terminated.  IF the LB deregistration delay runs beyond that, we
will end up with an instance that is terminated, but still registered
in the LB, which could lead to some dropped requests.

Some services have long running tasks or do not tolerate being
restarted frequently as well so those are pinned to not use spot
instances.

Closes SRE-4347