fix: use vanity url for static assets when CDN is disabled
The static asset root should be set to the vanity url when Cloudfront
is not in use.

Putting this behind the var.cloudfront_route_static_asset_traffic
flag was missed in a previous PR.  What can happen is that during
a blue/green deploy, static assets will 404 since session stickiness
will not be respected if the STATIC_ASSET_ROOT does not match
the main vanity url.