feat: datawatch encryption key resource policy
This change adds a resource policy to the Datawatch encryption key to ensure
that only the Datawatch role and additional specified principals are able
to read and write the value of the secret in Secrets Manager.

This behavior can be disabled by setting
`var.datawatch_encryption_key_restricted_policy = false`.